pyproject.toml
[tool.poetry]
name = "desafiols"
version = "0.1.0"
description = ""
authors = ["thiagoarvore <137798387+thiagoarvore@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
